feat: implement wave rendering functions and refactor useWaveRenderer hook
